refactor model into app
