--- .idea/modules/DataModel/RT.DataModel.iml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id=":DataModel" external.linked.project.path="$MODULE_DIR$/../../../DataModel" external.root.project.path="$MODULE_DIR$/../../.." external.system.id="GRADLE" external.system.module.group="RT" external.system.module.version="unspecified" type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$/../../../DataModel">
-      <excludeFolder url="file://$MODULE_DIR$/../../../DataModel/.gradle" />
-      <excludeFolder url="file://$MODULE_DIR$/../../../DataModel/build" />
-      <excludeFolder url="file://$MODULE_DIR$/../../../DataModel/out" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-</module>
