--- .idea/modules/DataModel/RT.DataModel.main.iml
@@ -1,68 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id=":DataModel:main" external.linked.project.path="$MODULE_DIR$/../../../DataModel" external.root.project.path="$MODULE_DIR$/../../.." external.system.id="GRADLE" external.system.module.group="RT" external.system.module.type="sourceSet" external.system.module.version="unspecified" type="JAVA_MODULE" version="4">
-  <component name="FacetManager">
-    <facet type="kotlin-language" name="Kotlin">
-      <configuration version="3" platform="JVM 1.8" useProjectSettings="false">
-        <compilerSettings />
-        <compilerArguments>
-          <option name="destination" value="$MODULE_DIR$/../../../DataModel/build/classes/kotlin/main" />
-          <option name="classpath" value="$MODULE_DIR$/../../../libs/TBA-APIv3-Java-Library-1.2.jar;C:/Users/Yu/IdeaProjects/Restructured-Tables/api/build/libs/ContextAPI.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/de.mpicbg.scicomp/krangl/0.10.3/c9ed5908f2d54afa4e3b883030a35b41bd61a6ed/krangl-0.10.3.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/com.beust/klaxon/3.0.1/808bd383de5140c185f11f5b38c75e6032076304/klaxon-3.0.1.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.jetbrains.kotlin/kotlin-reflect/1.3.11/aae7b33412715e9ed441934c4ffaad1bb80e9d36/kotlin-reflect-1.3.11.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/com.kyonifer/koma-core-ejml/0.12/29a23e1f9a7c71814f20f029f7b861abbd8e3282/koma-core-ejml-0.12.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/com.kyonifer/koma-core-api-jvm/0.12/c5650f155affe349a95c4d6c39e9292e32705852/koma-core-api-jvm-0.12.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.jetbrains.kotlin/kotlin-stdlib/1.3.11/4cbc5922a54376018307a731162ccaf3ef851a39/kotlin-stdlib-1.3.11.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/com.kyonifer/koma-core-api-common/0.12/8fbf821e7ff257aded8b9422edc1a0ce1f98e36a/koma-core-api-common-0.12.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.jetbrains.kotlin/kotlin-stdlib-common/1.3.11/d8b8e746e279f1c4f5e08bc14a96b82e6bb1de02/kotlin-stdlib-common-1.3.11.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.jetbrains/annotations/13.0/919f0dfe192fb4e063e7dacadee7f8bb9a2672a9/annotations-13.0.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.ejml/ejml-all/0.33/cace9897f75d225c7eefb69011c7b9119f4f3a9b/ejml-all-0.33.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.apache.commons/commons-csv/1.3/aeed8320e1b86b27e0b477a898eb7dd049526963/commons-csv-1.3.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.ejml/ejml-simple/0.33/73e00316092deb19d0fe174c718dcbe4bd75efe8/ejml-simple-0.33.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.ejml/ejml-fdense/0.33/449de4bf5e4f6fcbc204105db32bc8a56062ad4/ejml-fdense-0.33.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.ejml/ejml-dsparse/0.33/720ca88b50cc6aacbf48562ea01a8e8c6d72f9c/ejml-dsparse-0.33.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.ejml/ejml-ddense/0.33/8c578379904d16aefc6eedb278fc771df23aaabb/ejml-ddense-0.33.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.ejml/ejml-cdense/0.33/4a47dd19a4fdb4df5b3d38b34672b0f103b1d64d/ejml-cdense-0.33.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.ejml/ejml-zdense/0.33/b9392048ff7d52ca090c3efdbc16fa56de737a00/ejml-zdense-0.33.jar;C:/Users/Yu/.gradle/caches/modules-2/files-2.1/org.ejml/ejml-core/0.33/79f1e14858b9774bb169d4e40b2a07aa6950bb62/ejml-core-0.33.jar" />
-          <option name="noStdlib" value="true" />
-          <option name="noReflect" value="true" />
-          <option name="moduleName" value="DataModel" />
-          <option name="jvmTarget" value="1.8" />
-          <option name="addCompilerBuiltIns" value="true" />
-          <option name="loadBuiltInsFromDependencies" value="true" />
-          <option name="languageVersion" value="1.3" />
-          <option name="apiVersion" value="1.3" />
-          <option name="pluginOptions">
-            <array />
-          </option>
-          <option name="pluginClasspaths">
-            <array />
-          </option>
-          <option name="errors">
-            <ArgumentParseErrors />
-          </option>
-        </compilerArguments>
-      </configuration>
-    </facet>
-  </component>
-  <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_8">
-    <output url="file://$MODULE_DIR$/../../../DataModel/out/production/classes" />
-    <exclude-output />
-    <content url="file://$MODULE_DIR$/../../../DataModel/src/main">
-      <sourceFolder url="file://$MODULE_DIR$/../../../DataModel/src/main/kotlin" isTestSource="false" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../libs/TBA-APIv3-Java-Library-1.2.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module" module-name="RT.ContextAPI.main" />
-    <orderEntry type="library" name="Gradle: de.mpicbg.scicomp:krangl:0.10.3" level="project" />
-    <orderEntry type="library" name="Gradle: com.beust:klaxon:3.0.1" level="project" />
-    <orderEntry type="library" name="Gradle: org.jetbrains.kotlin:kotlin-reflect:1.3.11" level="project" />
-    <orderEntry type="library" name="Gradle: com.kyonifer:koma-core-ejml:0.12" level="project" />
-    <orderEntry type="library" name="Gradle: com.kyonifer:koma-core-api-jvm:0.12" level="project" />
-    <orderEntry type="library" name="Gradle: org.jetbrains.kotlin:kotlin-stdlib:1.3.11" level="project" />
-    <orderEntry type="library" name="Gradle: com.kyonifer:koma-core-api-common:0.12" level="project" />
-    <orderEntry type="library" name="Gradle: org.jetbrains.kotlin:kotlin-stdlib-common:1.3.11" level="project" />
-    <orderEntry type="library" name="Gradle: org.jetbrains:annotations:13.0" level="project" />
-    <orderEntry type="library" name="Gradle: org.ejml:ejml-all:0.33" level="project" />
-    <orderEntry type="library" name="Gradle: org.apache.commons:commons-csv:1.3" level="project" />
-    <orderEntry type="library" name="Gradle: org.ejml:ejml-simple:0.33" level="project" />
-    <orderEntry type="library" name="Gradle: org.ejml:ejml-fdense:0.33" level="project" />
-    <orderEntry type="library" name="Gradle: org.ejml:ejml-dsparse:0.33" level="project" />
-    <orderEntry type="library" name="Gradle: org.ejml:ejml-ddense:0.33" level="project" />
-    <orderEntry type="library" name="Gradle: org.ejml:ejml-cdense:0.33" level="project" />
-    <orderEntry type="library" name="Gradle: org.ejml:ejml-zdense:0.33" level="project" />
-    <orderEntry type="library" name="Gradle: org.ejml:ejml-core:0.33" level="project" />
-  </component>
-</module>
