+++ app/src/main/kotlin/ca/warp7/rt/core/model/AutoTypes.kt
@@ -1,6 +1,6 @@
 @file:Suppress("unused")
 
-package ca.warp7.rt.context.model
+package ca.warp7.rt.core.model
 
 import ca.warp7.rt.context.api.*
