+++ app/src/main/kotlin/ca/warp7/rt/core/model/MetricsImpl.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MemberVisibilityCanBePrivate", "unused")
 
-package ca.warp7.rt.context.model
+package ca.warp7.rt.core.model
 
 import ca.warp7.rt.context.api.IntMetric
 import ca.warp7.rt.context.api.PipelineStreamVector
