+++ settings.gradle.kts
@@ -1,7 +1,6 @@
 rootProject.name = "RT"
 
 include(":api")
-include(":DataModel")
 include(":app")
 
 project(":api").name = "ContextAPI"
